power: regulator: Perform regulator setup inside uclass
Regulators initial setup was previously dependent on board call.
To move from this behaviour next solution is proposed: on post_bind
boot-on/always-on properties are checked, all regulators with
such props will be probed just after binding which ensures that
essential regulators are set, then in the post probe regulator
autoset is called so that correct regulator state according to
device tree is reached.

Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>
[jonas@kwiboo.se: use autoset func, only probe with always/boot-on prop]
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>